Feat mesh rendering (#142)
* Changed editor info and editor camera component fetches to use references. Fixed mouseInScene bug

* Started mesh rendering refactor

* Formatted

* Fixed NDC mapping

* Refactored more of the material system, aided by AI

* Refactored grid and mesh material system to ensure proper ordering, this commit was made with the assistance of AI

* Created dynamic buffers by resizing

* Feat buffer resizing (#143)

* Created dynamic buffers by resizing

* Fixed memory read bounds and formatting

* Fixed comments on PR

* Formatted

* Feat mesh textures (#145)

* Created dynamic buffers by resizing

* Fixed memory read bounds and formatting

* Fixed comments on PR

* Formatted

* Meshes working with textures

* Use normals by default on GLTF load and correct directional light usage

* Formatted files

* Fixed some comments

* Moved editor camera to comp ref

* Fixed most of the comments

* Formatted

* Merged with dev and fixed issues

* Updated CMakeLists

* Removed unnecesary code failing on WASM

* Fixed size mismatch on WASM

* Fixed warnings for wasm

+++ src/engine_core/resources/src/Components/GpuUploadComponent.hpp
@@ -6,7 +6,7 @@
 
 #pragma once
 
-#include <cstddef>
+#include <cstdint>
 
 namespace Hush::Renderer
 {
@@ -32,7 +32,7 @@ namespace Hush::Renderer
 					  ///< quota.
 		};
 
-		size_t cpuOffset = 0; ///< Byte offset into the staging buffer where this entity's data is located.
+		uint64_t cpuOffset = 0; ///< Byte offset into the staging buffer where this entity's data is located.
 
 		EUploadMode uploadMode = EUploadMode::Full; ///< Upload mode for this entity's resource.
 	};
